+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/app.py
@@ -561,6 +561,8 @@ class AppLauncher:
         # add orbit modules back to sys.modules
         for key, value in hacked_modules.items():
             sys.modules[key] = value
+        # hide the stop button in the toolbar
+        self._hide_stop_button()
 
     def _load_extensions(self):
         """Load correct extensions based on AppLauncher's resolved config member variables."""
@@ -644,3 +646,15 @@ class AppLauncher:
                 "/persistent/isaac/asset_root/nvidia",
                 "http://omniverse-content-production.s3-us-west-2.amazonaws.com/Assets",
             )
+
+    def _hide_stop_button(self):
+        """Hide the stop button in the toolbar."""
+        import omni.kit.widget.toolbar
+
+        # grey out the stop button because we don't want to stop the simulation manually in standalone mode
+        toolbar = omni.kit.widget.toolbar.get_instance()
+        play_button_group = toolbar._builtin_tools._play_button_group  # type: ignore
+        if play_button_group is not None:
+            play_button_group._stop_button.visible = False  # type: ignore
+            play_button_group._stop_button.enabled = False  # type: ignore
+            play_button_group._stop_button = None  # type: ignore