+++ source/extensions/omni.isaac.orbit_tasks/test/wrappers/test_sb3_wrapper.py
@@ -66,8 +66,6 @@ class TestStableBaselines3VecEnvWrapper(unittest.TestCase):
             omni.usd.get_context().new_stage()
             # parse configuration
             env_cfg: RLTaskEnvCfg = parse_env_cfg(task_name, use_gpu=self.use_gpu, num_envs=self.num_envs)
-            # note: we don't want to shutdown the app on stop during the tests since we reload the stage
-            env_cfg.sim.shutdown_app_on_stop = False
 
             # create environment
             env = gym.make(task_name, cfg=env_cfg)